PET-120 Refactor: entity 패키지 변경
